Fixed if statements etc

+++ partials/blog_posts.php
@@ -6,12 +6,16 @@
     
             <article>
                 <?php
-                if ($article['user_id'] == $_SESSION['id']):
+                if(isset($_SESSION['signed_in'])):
+                    if ($article['user_id'] == $_SESSION['id']):
+                        ?>
+                        <a href="partials/edit_blogpost.php?id=<?= $article['postID']?>&content=<?=$article['text'] ?>">Edit your blog post here</a> 
+                        <br/>
+                        <a href="partials/delete_blogpost.php?id=<?= $article['postID']?>">Delete blog post</a>
+                    <?php 
+                    endif;
+                endif;
                     ?>
-                    <a href="partials/edit_blogpost.php?id=<?= $article['postID']?>&content=<?=$article['text'] ?>">Edit your blog post here</a> 
-                    <br/>
-                    <a href="partials/delete_blogpost.php?id=<?= $article['postID']?>">Delete blog post</a>
-                <?php endif; ?>
 
                 <h2><?= $article['post_title']; ?></h2>
